+++ CHANGELOG.md
@@ -8,6 +8,12 @@ vNext
 
 Upcoming changes for the next release.
 
+### Features
+
+- Adds selected option render prop on Select component to override default input value format.
+  Adds Stepper component, theme, skin, and stories.
+  [PR 115](https://github.com/input-output-hk/react-polymorph/pull/115)
+
 0.8.6
 =====